Upgrade file pointer buffers from int to unsigned long int

+++ FTL.h
@@ -170,7 +170,7 @@ overTimeDataStruct *overTime;
 
 FILE *logfile;
 FILE *dnsmasqlog;
-int dnsmasqlogpos;
+unsigned long int dnsmasqlogpos;
 volatile sig_atomic_t killed;
 
 char ** setupVarsArray;
